Objective
To find the player of the series and man of the match award winner in a
season. 

To find the team that won the most number of matches in a season. 

Does winning toss increases the chances of victory with first bat and first
 field. 

To find the most winning team for each season.